[Impeller] move aiks text tests to DL.
