+++ impeller/aiks/aiks_unittests.cc
@@ -33,9 +33,6 @@
 #include "impeller/renderer/command_buffer.h"
 #include "impeller/renderer/snapshot.h"
 #include "impeller/typographer/backends/skia/text_frame_skia.h"
-#include "impeller/typographer/backends/stb/text_frame_stb.h"
-#include "impeller/typographer/backends/stb/typeface_stb.h"
-#include "impeller/typographer/backends/stb/typographer_context_stb.h"
 #include "third_party/imgui/imgui.h"
 #include "third_party/skia/include/core/SkFontMgr.h"
 #include "txt/platform.h"
@@ -203,352 +200,6 @@ TEST_P(AiksTest, CanPerformSaveLayerWithBounds) {
   canvas.Restore();
 
   ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest,
-       CanPerformSaveLayerWithBoundsAndLargerIntermediateIsNotAllocated) {
-  Canvas canvas;
-
-  Paint red;
-  red.color = Color::Red();
-
-  Paint green;
-  green.color = Color::Green();
-
-  Paint blue;
-  blue.color = Color::Blue();
-
-  Paint save;
-  save.color = Color::Black().WithAlpha(0.5);
-
-  canvas.SaveLayer(save, Rect::MakeXYWH(0, 0, 100000, 100000));
-
-  canvas.DrawRect(Rect::MakeXYWH(0, 0, 100, 100), red);
-  canvas.DrawRect(Rect::MakeXYWH(10, 10, 100, 100), green);
-  canvas.DrawRect(Rect::MakeXYWH(20, 20, 100, 100), blue);
-
-  canvas.Restore();
-
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-struct TextRenderOptions {
-  bool stroke = false;
-  Scalar font_size = 50;
-  Color color = Color::Yellow();
-  Point position = Vector2(100, 200);
-  std::optional<Paint::MaskBlurDescriptor> mask_blur_descriptor;
-};
-
-bool RenderTextInCanvasSkia(const std::shared_ptr<Context>& context,
-                            Canvas& canvas,
-                            const std::string& text,
-                            const std::string_view& font_fixture,
-                            TextRenderOptions options = {}) {
-  // Draw the baseline.
-  canvas.DrawRect(
-      Rect::MakeXYWH(options.position.x - 50, options.position.y, 900, 10),
-      Paint{.color = Color::Aqua().WithAlpha(0.25)});
-
-  // Mark the point at which the text is drawn.
-  canvas.DrawCircle(options.position, 5.0,
-                    Paint{.color = Color::Red().WithAlpha(0.25)});
-
-  // Construct the text blob.
-  auto c_font_fixture = std::string(font_fixture);
-  auto mapping = flutter::testing::OpenFixtureAsSkData(c_font_fixture.c_str());
-  if (!mapping) {
-    return false;
-  }
-  sk_sp<SkFontMgr> font_mgr = txt::GetDefaultFontManager();
-  SkFont sk_font(font_mgr->makeFromData(mapping), options.font_size);
-  auto blob = SkTextBlob::MakeFromString(text.c_str(), sk_font);
-  if (!blob) {
-    return false;
-  }
-
-  // Create the Impeller text frame and draw it at the designated baseline.
-  auto frame = MakeTextFrameFromTextBlobSkia(blob);
-
-  Paint text_paint;
-  text_paint.color = options.color;
-  text_paint.mask_blur_descriptor = options.mask_blur_descriptor;
-  text_paint.stroke_width = 1;
-  text_paint.style =
-      options.stroke ? Paint::Style::kStroke : Paint::Style::kFill;
-  canvas.DrawTextFrame(frame, options.position, text_paint);
-  return true;
-}
-
-bool RenderTextInCanvasSTB(const std::shared_ptr<Context>& context,
-                           Canvas& canvas,
-                           const std::string& text,
-                           const std::string& font_fixture,
-                           TextRenderOptions options = {}) {
-  // Draw the baseline.
-  canvas.DrawRect(
-      Rect::MakeXYWH(options.position.x - 50, options.position.y, 900, 10),
-      Paint{.color = Color::Aqua().WithAlpha(0.25)});
-
-  // Mark the point at which the text is drawn.
-  canvas.DrawCircle(options.position, 5.0,
-                    Paint{.color = Color::Red().WithAlpha(0.25)});
-
-  // Construct the text blob.
-  auto mapping = flutter::testing::OpenFixtureAsMapping(font_fixture.c_str());
-  if (!mapping) {
-    return false;
-  }
-  auto typeface_stb = std::make_shared<TypefaceSTB>(std::move(mapping));
-
-  auto frame = MakeTextFrameSTB(
-      typeface_stb, Font::Metrics{.point_size = options.font_size}, text);
-
-  Paint text_paint;
-  text_paint.color = options.color;
-  canvas.DrawTextFrame(frame, options.position, text_paint);
-  return true;
-}
-
-TEST_P(AiksTest, CanRenderTextFrame) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextFrameWithInvertedTransform) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  canvas.Translate({1000, 0, 0});
-  canvas.Scale({-1, 1, 1});
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderStrokedTextFrame) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf",
-      {
-          .stroke = true,
-      }));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextFrameWithHalfScaling) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-  canvas.Scale({0.5, 0.5, 1});
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextFrameWithFractionScaling) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-  canvas.Scale({2.625, 2.625, 1});
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextFrameSTB) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-  ASSERT_TRUE(RenderTextInCanvasSTB(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-
-  SetTypographerContext(TypographerContextSTB::Make());
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, TextFrameSubpixelAlignment) {
-  // "Random" numbers between 0 and 1. Hardcoded to avoid flakiness in goldens.
-  std::array<Scalar, 20> phase_offsets = {
-      7.82637e-06, 0.131538,  0.755605,   0.45865,   0.532767,
-      0.218959,    0.0470446, 0.678865,   0.679296,  0.934693,
-      0.383502,    0.519416,  0.830965,   0.0345721, 0.0534616,
-      0.5297,      0.671149,  0.00769819, 0.383416,  0.0668422};
-  auto callback = [&](AiksContext& renderer) -> std::optional<Picture> {
-    static float font_size = 20;
-    static float phase_variation = 0.2;
-    static float speed = 0.5;
-    static float magnitude = 100;
-    if (AiksTest::ImGuiBegin("Controls", nullptr,
-                             ImGuiWindowFlags_AlwaysAutoResize)) {
-      ImGui::SliderFloat("Font size", &font_size, 5, 50);
-      ImGui::SliderFloat("Phase variation", &phase_variation, 0, 1);
-      ImGui::SliderFloat("Oscillation speed", &speed, 0, 2);
-      ImGui::SliderFloat("Oscillation magnitude", &magnitude, 0, 300);
-      ImGui::End();
-    }
-
-    Canvas canvas;
-    canvas.Scale(GetContentScale());
-
-    for (size_t i = 0; i < phase_offsets.size(); i++) {
-      auto position =
-          Point(200 + magnitude *
-                          std::sin((-phase_offsets[i] * k2Pi * phase_variation +
-                                    GetSecondsElapsed() * speed)),  //
-                200 + i * font_size * 1.1                           //
-          );
-      if (!RenderTextInCanvasSkia(
-              GetContext(), canvas,
-              "the quick brown fox jumped over "
-              "the lazy dog!.?",
-              "Roboto-Regular.ttf",
-              {.font_size = font_size, .position = position})) {
-        return std::nullopt;
-      }
-    }
-    return canvas.EndRecordingAsPicture();
-  };
-
-  ASSERT_TRUE(OpenPlaygroundHere(callback));
-}
-
-TEST_P(AiksTest, CanRenderItalicizedText) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "HomemadeApple.ttf"));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-static constexpr std::string_view kFontFixture =
-#if FML_OS_MACOSX
-    "Apple Color Emoji.ttc";
-#else
-    "NotoColorEmoji.ttf";
-#endif
-
-TEST_P(AiksTest, CanRenderEmojiTextFrame) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "😀 😃 😄 😁 😆 😅 😂 🤣 🥲 😊", kFontFixture));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderEmojiTextFrameWithBlur) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "😀 😃 😄 😁 😆 😅 😂 🤣 🥲 😊", kFontFixture,
-      TextRenderOptions{.color = Color::Blue(),
-                        .mask_blur_descriptor = Paint::MaskBlurDescriptor{
-                            .style = FilterContents::BlurStyle::kNormal,
-                            .sigma = Sigma(4)}}));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderEmojiTextFrameWithAlpha) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "😀 😃 😄 😁 😆 😅 😂 🤣 🥲 😊", kFontFixture,
-      {.color = Color::Black().WithAlpha(0.5)}));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextInSaveLayer) {
-  Canvas canvas;
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  canvas.Translate({100, 100});
-  canvas.Scale(Vector2{0.5, 0.5});
-
-  // Blend the layer with the parent pass using kClear to expose the coverage.
-  canvas.SaveLayer({.blend_mode = BlendMode::kClear});
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-  canvas.Restore();
-
-  // Render the text again over the cleared coverage rect.
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextOutsideBoundaries) {
-  Canvas canvas;
-  canvas.Translate({200, 150});
-
-  // Construct the text blob.
-  auto mapping = flutter::testing::OpenFixtureAsSkData("wtf.otf");
-  ASSERT_NE(mapping, nullptr);
-
-  Scalar font_size = 80;
-  sk_sp<SkFontMgr> font_mgr = txt::GetDefaultFontManager();
-  SkFont sk_font(font_mgr->makeFromData(mapping), font_size);
-
-  Paint text_paint;
-  text_paint.color = Color::Blue().WithAlpha(0.8);
-
-  struct {
-    Point position;
-    const char* text;
-  } text[] = {{Point(0, 0), "0F0F0F0"},
-              {Point(1, 2), "789"},
-              {Point(1, 3), "456"},
-              {Point(1, 4), "123"},
-              {Point(0, 6), "0F0F0F0"}};
-  for (auto& t : text) {
-    canvas.Save();
-    canvas.Translate(t.position * Point(font_size * 2, font_size * 1.1));
-    {
-      auto blob = SkTextBlob::MakeFromString(t.text, sk_font);
-      ASSERT_NE(blob, nullptr);
-      auto frame = MakeTextFrameFromTextBlobSkia(blob);
-      canvas.DrawTextFrame(frame, Point(), text_paint);
-    }
-    canvas.Restore();
-  }
-
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, TextRotated) {
-  Canvas canvas;
-  canvas.Scale(GetContentScale());
-  canvas.DrawPaint({.color = Color(0.1, 0.1, 0.1, 1.0)});
-
-  canvas.Transform(Matrix(0.25, -0.3, 0, -0.002,  //
-                          0, 0.5, 0, 0,           //
-                          0, 0, 0.3, 0,           //
-                          100, 100, 0, 1.3));
-  ASSERT_TRUE(RenderTextInCanvasSkia(
-      GetContext(), canvas, "the quick brown fox jumped over the lazy dog!.?",
-      "Roboto-Regular.ttf"));
-
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-// This makes sure the WideGamut named tests use 16bit float pixel format.
-TEST_P(AiksTest, FormatWideGamut) {
-  EXPECT_EQ(GetContext()->GetCapabilities()->GetDefaultColorFormat(),
-            PixelFormat::kB10G10R10A10XR);
 }
 
 TEST_P(AiksTest, FormatSRGB) {
@@ -1326,35 +977,6 @@ TEST_P(AiksTest, SolidColorApplyColorFilter) {
                     Color(0.424452, 0.828743, 0.79105, 0.9375));
 }
 
-TEST_P(AiksTest, DrawScaledTextWithPerspectiveNoSaveLayer) {
-  Canvas canvas;
-  canvas.Transform(Matrix(1.0, 0.0, 0.0, 0.0,    //
-                          0.0, 1.0, 0.0, 0.0,    //
-                          0.0, 0.0, 1.0, 0.01,   //
-                          0.0, 0.0, 0.0, 1.0) *  //
-                   Matrix::MakeRotationY({Degrees{10}}));
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(GetContext(), canvas, "Hello world",
-                                     "Roboto-Regular.ttf"));
-
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, DrawScaledTextWithPerspectiveSaveLayer) {
-  Canvas canvas;
-  Paint save_paint;
-  canvas.SaveLayer(save_paint);
-  canvas.Transform(Matrix(1.0, 0.0, 0.0, 0.0,    //
-                          0.0, 1.0, 0.0, 0.0,    //
-                          0.0, 0.0, 1.0, 0.01,   //
-                          0.0, 0.0, 0.0, 1.0) *  //
-                   Matrix::MakeRotationY({Degrees{10}}));
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(GetContext(), canvas, "Hello world",
-                                     "Roboto-Regular.ttf"));
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
 TEST_P(AiksTest, PipelineBlendSingleParameter) {
   Canvas canvas;
 
@@ -1556,24 +1178,6 @@ TEST_P(AiksTest, StrokedPathWithMoveToThenCloseDrawnCorrectly) {
                             .stroke_cap = Cap::kRound,
                             .style = Paint::Style::kStroke,
                         });
-  ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
-}
-
-TEST_P(AiksTest, CanRenderTextWithLargePerspectiveTransform) {
-  // Verifies that text scales are clamped to work around
-  // https://github.com/flutter/flutter/issues/136112 .
-
-  Canvas canvas;
-  Paint save_paint;
-  canvas.SaveLayer(save_paint);
-  canvas.Transform(Matrix(2000, 0, 0, 0,   //
-                          0, 2000, 0, 0,   //
-                          0, 0, -1, 9000,  //
-                          0, 0, -1, 7000   //
-                          ));
-
-  ASSERT_TRUE(RenderTextInCanvasSkia(GetContext(), canvas, "Hello world",
-                                     "Roboto-Regular.ttf"));
   ASSERT_TRUE(OpenPlaygroundHere(canvas.EndRecordingAsPicture()));
 }
