+++ impeller/display_list/BUILD.gn
@@ -59,6 +59,7 @@ template("display_list_unittests_component") {
     "aiks_dl_opacity_unittests.cc",
     "aiks_dl_path_unittests.cc",
     "aiks_dl_runtime_effect_unittests.cc",
+    "aiks_dl_text_unittests.cc",
     "aiks_dl_unittests.cc",
     "aiks_dl_vertices_unittests.cc",
     "dl_golden_blur_unittests.cc",
